Added miner search functionality
Adjusted some sorting layers.
Now generates a miner that is placed in a random cave at the max depth. If an incomplete tree doesn't fill that fully, it will search the next level up until it finds one.
Player and Miner cave indices are now tracked.
Player can now press "S" to shout for the miner. If the miner is in the left subtree, the debug log will say he is on the left, and "I'm over here!" will show on the bottom left of the cave. Directions and text placement change if he is in the right subtree, or behind if the Player made the wrong choice.
On entering the correct cave, the debug log will say the Player found the miner. This also sets a boolean value found in the Player's script to true.