Smaller Off-Center Miner & Bug Fix
The lost miner is now off to the side of the cave and much smaller. This is easy to change back, but it's an idea.

Bug Fix - The player visiting a leaf node neighboring the one with the lost miner no longer triggers the "Got to the miner!" message. I just made the collider really small, so the event only triggers when the player gets to the dead center of the cave.

The trouble with a medium-size collider is that if you click directly on it, the player doesn't move. That sometimes happens with a larger collider too, however.