Camera fixes and more
The camera now has a 3rd zoom level, moving in on the node entirely, allowing for easier debris clearing and seeing details in larger trees. It is much smoother when transitioning, with the worst of it being on use in the first second or so.

The cave exit sprites now have a collider, which is narrow along the center of the exit and extends partially over the tunnel. This allows easy movement even when zoomed in entirely on the node.

The player now moves more smoothly from node to node, and for extremely large routes he is sped up drastically. I had to change the code a bit more to make that work with the camera, but I don't think you guys worked on that.

I updated sprites for the miner shouts and thanks, and their size works better with the newer cave and miner sprites. They are also positioned along the tunnels, reducing potential confusion.